backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
google-generativeai
psycopg2-binary
